Crate fastcrypto

Re-exports

pub use signature::Signature as _;

Modules

bls12381

bulletproofs

ed25519

hkdf

pubkey_bytes

secp256k1

serde_helpers

traits

Structs

Digest

Represents a hash digest (32 bytes).

SignatureService

This service holds the node’s private key. It takes digests as input and returns a signature over the digest (through a one-shot channel).

Constants

DIGEST_LEN

Traits

Hash

This trait is implemented by all messages that can be hashed.

Verifier

Verify the provided message bytestring using Self (e.g. a public key)

Functions

blake2b_256

generate_keypair

generate_production_keypair