fastcrypto::aes

Trait Cipher

pub trait Cipher {
    type IVType: Nonce;

    // Required methods
    fn encrypt(&self, iv: &Self::IVType, plaintext: &[u8]) -> Vec<u8> ;
    fn decrypt(
        &self,
        iv: &Self::IVType,
        ciphertext: &[u8],
    ) -> Result<Vec<u8>, FastCryptoError>;
}

Trait impl’d by symmetric ciphers.

Required Associated Types

type IVType: Nonce

Required Methods

fn encrypt(&self, iv: &Self::IVType, plaintext: &[u8]) -> Vec<u8>

Encrypt plaintext using the given IV and return the result.

fn decrypt( &self, iv: &Self::IVType, ciphertext: &[u8], ) -> Result<Vec<u8>, FastCryptoError>

Decrypt ciphertext using the given IV and return the result. An error may be returned in CBC-mode if the ciphertext is not correctly padded, but in other modes this method always return Ok.