Module fastcrypto::encoding

Encodings of binary data such as Base64 and Hex.

Example

assert_eq!(Hex::encode("Hello world!"), "0x48656c6c6f20776f726c6421");
assert_eq!(Base64::encode("Hello world!"), "SGVsbG8gd29ybGQh");

Structs

Base64

Base64 encoding.

Hex

Hex string encoding.

Traits

Encoding

Trait representing a general binary-to-string encoding.

Functions

decode_bytes_hex

Decodes a hex string to bytes. Both upper and lower case characters are allowed in the hex string.

encode_bytes_hex

Encodes bytes as a hex string using lower case characters.