fastcrypto

Module error

Collection of errors to be used in fastcrypto.

A function should validate its arguments and return an indicative error where needed. However, once the function is executing the cryptographic protocol/algorithm (directly/ indirectly) then it should not return explicit errors as it might leak private information. In those cases the function should return the opaque, general error FastCryptoError::GeneralOpaqueError. When in doubt, prefer FastCryptoError::GeneralOpaqueError.

Enums

FastCryptoError

Collection of errors to be used in fastcrypto.

Type Aliases

FastCryptoResult