Trait fastcrypto::groups::multiplier::ScalarMultiplier

pub trait ScalarMultiplier<G: GroupElement> {
    // Required methods
    fn new(base_element: G) -> Self;
    fn mul(&self, scalar: &G::ScalarType) -> G;

    // Provided method
    fn two_scalar_mul(
        &self,
        base_scalar: &G::ScalarType,
        other_element: &G,
        other_scalar: &G::ScalarType
    ) -> G { ... }
}

Trait for scalar multiplication for a fixed group element, e.g. by using precomputed values.

Required Methods

fn new(base_element: G) -> Self

Create a new scalar multiplier with the given base element.

fn mul(&self, scalar: &G::ScalarType) -> G

Compute self.base_element * scalar.

Provided Methods

fn two_scalar_mul( &self, base_scalar: &G::ScalarType, other_element: &G, other_scalar: &G::ScalarType ) -> G

Compute self.base_element * base_scalar + other_element * other_scalar.