fastcrypto::groups::multiplier::windowed
Function multi_scalar_mul

pub fn multi_scalar_mul<G: Doubling + for<'a> Add<&'a G, Output = G> + Clone + Debug, S: ToLittleEndianBytes + Clone + Debug, const N: usize>(
    scalars: &[S; N],
    elements: &[G; N],
    precomputed_multiples: &HashMap<usize, Vec<G>>,
    default_window_width: usize,
    zero: G,
) -> G

This method computes the linear combination of the given scalars and group elements using the sliding window method. Some group elements may have tables of precomputed elements which can be given in the precomputed hash map. For the elements which does not have a precomputed table a table of size 2default_window_width - 1 is computed.

The precomputed tables for an element g should contain the multiples 2w-1 g , …, (2w - 1) g for some integer w > 1 which is the window width for the given element.

The default_window_width is the window width for the elements that does not have a precomputation table and may be set to any value >= 1. As rule-of-thumb, this should be set to approximately the bit length of the square root of the scalar size for optimal performance.