fastcrypto::groups::secp256k1

Struct ProjectivePoint

pub struct ProjectivePoint(/* private fields */);

A point on the Secp256k1 curve in projective coordinates.

Implementations

impl ProjectivePoint

pub fn x_as_be_bytes(&self) -> FastCryptoResult<[u8; 32]>

Returns the x-coordinate of this point as a big-endian byte array. Returns an InvalidInput error if this is the identity point.

pub fn with_even_y_from_x_be_bytes(x: &[u8; 32]) -> FastCryptoResult<Self>

Lift an x-coordinate to a point on the curve with an even y-coordinate. Returns an error if x is not the x-coordinate of a point on the curve.

pub fn has_even_y(&self) -> FastCryptoResult<bool>

Returns true iff the y-coordinate of this point is even. Returns an InvalidInput error if this is the identity point.

pub fn is_zero(&self) -> bool

Returns true iff this is the identity point aka the point at infinity.