fastcrypto::groups

Trait FiatShamirChallenge

pub trait FiatShamirChallenge {
    // Required method
    fn fiat_shamir_reduction_to_group_element(uniform_buffer: &[u8]) -> Self;
}

Trait for groups that have a reduction from a random buffer to a group element that is secure when used for Fiat-Shamir. Note that the resulting group element is not guaranteed to be uniformly distributed, but only to have enough entropy to be used for Fiat-Shamir heuristic.

Required Methods

fn fiat_shamir_reduction_to_group_element(uniform_buffer: &[u8]) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.