Trait fastcrypto::groups::GroupElement

pub trait GroupElement: Eq + Add + Sub + Neg + AddAssign + SubAssign + Mul<Self::ScalarType> + Sized {
    type ScalarType: Scalar;

    fn zero() -> Self;
    fn generator() -> Self;
}

Trait impl’d by elements of an additive cyclic group.

Required Associated Types

type ScalarType: Scalar

Type of scalars used in the [Self::mul] multiplication method.

Required Methods

fn zero() -> Self

Return an instance of the identity element in this group.

fn generator() -> Self

Return an instance of the generator for this group.