fastcrypto::hash

Trait Hash

pub trait Hash<const DIGEST_LEN: usize> {
    type TypedDigest: Into<Digest<DIGEST_LEN>> + Eq + Hash + Copy + Debug;

    // Required method
    fn digest(&self) -> Self::TypedDigest;
}

This trait is implemented by all messages that can be hashed.

Required Associated Types

type TypedDigest: Into<Digest<DIGEST_LEN>> + Eq + Hash + Copy + Debug

The type of the digest when this is hashed.

Required Methods

fn digest(&self) -> Self::TypedDigest