Trait fastcrypto::hash::ReverseWrapper

pub trait ReverseWrapper {
    type Variant: Digest + 'static + CoreProxy + OutputSizeUser;
}

This allows us to use the hash functions defined here and use them in place of their internal functions from the digest crate.

Required Associated Types

type Variant: Digest + 'static + CoreProxy + OutputSizeUser