Module fastcrypto::hmac

Implementations of HMAC and HKDF.

Functions

hkdf_generate_from_ikm

Creation of a keypair using the RFC 5869 HKDF specification. This requires choosing an HMAC function of the correct length (conservatively, the size of a private key for this curve). Despite the unsightly generics (which aim to ensure this works for a wide range of hash functions), this is straightforward to use.

hkdf_sha3_256

HMAC-based Extract-and-Expand Key Derivation Function (HKDF) using SHA3-256.

hmac_sha3_256

Keyed-Hash Message Authentication Code (HMAC) using SHA3-256.

Type Definitions

HkdfIkm

Type for input keying material in hkdf_sha3_256.

HmacKey

Type for key in hmac_sha3_256.