fastcrypto

Module nitro_attestation

Verification of AWS Nitro enclave attestation documents: COSE_Sign1 parsing, attestation document validation and certificate chain verification against the AWS Nitro root of trust, per https://docs.aws.amazon.com/enclaves/latest/user/verify-root.html.

Structs

AttestationDocument

The AWS Nitro Attestation Document, see https://docs.aws.amazon.com/enclaves/latest/user/verify-root.html#doc-def

CoseSign1

Implementation of the COSE_Sign1 structure as defined in RFC8152. protected_header: See Section 3 (Note: AWS Nitro does not have unprotected header.) payload: See Section 4.2. signature: See Section 4.2. Class and trait impl adapted from https://github.com/awslabs/aws-nitro-enclaves-cose/blob/main/src/sign.rs

HeaderMap

Empty map wrapper for COSE headers.

Enums

NitroAttestationVerifyError

Error type for Nitro attestation verification.

Functions

parse_nitro_attestation

Parse an attestation document into (signature, signed_message, payload).

verify_nitro_attestation

Given the signature bytes, signed message and parsed payload, verify everything according to https://docs.aws.amazon.com/enclaves/latest/user/verify-root.html and https://github.com/aws/aws-nitro-enclaves-nsm-api/blob/main/docs/attestation_process.md.