Module fastcrypto::secp256k1::recoverable

This module contains an implementation of the ECDSA signature scheme over the secp256k1 curve.

Messages can be signed and the public key can be recovered from the signature:

Example

use rand::thread_rng;
let kp = Secp256k1KeyPair::generate(&mut thread_rng());
let message: &[u8] = b"Hello, world!";
let signature = kp.sign_recoverable(message);
assert_eq!(&signature.recover(message).unwrap(), kp.public());

Structs

Secp256k1RecoverableSignature

Secp256k1 signature.

Secp256k1Sig

An ECDSA signature

Constants

SECP256K1_RECOVERABLE_SIGNATURE_SIZE

Length of a compact signature followed by one extra byte for recovery id, used to recover the public key from a signature.

Statics

SECP256K1