fastcrypto::secp256r1
Module recoverable

This module contains an implementation of the ECDSA signature scheme over the secp256r1 NIST-P1 curve where the public key can be recovered from a signature. The nonce is generated deterministically according to RFC6979.

Messages can be signed and the public key can be recovered from the signature:

Example

use rand::thread_rng;
let kp = Secp256r1KeyPair::generate(&mut thread_rng());
let message: &[u8] = b"Hello, world!";
let signature = kp.sign_recoverable(message);
assert_eq!(kp.public(), &signature.recover(message).unwrap());

Structs

Secp256r1RecoverableSignature

Secp256r1 signature.

Constants

SECP256R1_RECOVERABLE_SIGNATURE_LENGTH