Struct fastcrypto::secp256r1::recoverable::Secp256r1RecoverableSignature

pub struct Secp256r1RecoverableSignature {
    pub sig: ExternalSignature,
    pub bytes: OnceCell<[u8; 65]>,
    pub recovery_id: u8,
}

Secp256r1 signature.

Fields
sig: ExternalSignaturebytes: OnceCell<[u8; 65]>recovery_id: u8
Implementations

impl Secp256r1RecoverableSignature

pub fn try_from_nonrecoverable( signature: &Secp256r1Signature, pk: &Secp256r1PublicKey, message: &[u8] ) -> Result<Self, FastCryptoError>

Recover the public key used to create this signature. This assumes the recovery id byte has been set. The hash function H is used to hash the message.

An FastCryptoError::GeneralOpaqueError is returned if no public keys can be recovered.