Struct fastcrypto::serde_helpers::BytesRepresentation

pub struct BytesRepresentation<const N: usize>(pub [u8; N]);

External types.

BytesRepresentation is a basic wrapper for storing bincode serialized objects.

To be used in external interfaces instead of the internal types.

Derive by calling generate_bytes_representation.

Uses Base64 when serialized with a human readable serializer, and raw bytes otherwise.

Note that in theory external types should not be stored in Sui and only be serialized with is_human_readable(). For storage and computation in Sui, internal types should be used (see above).

Tuple Fields
0: [u8; N]