Module fastcrypto::traits

Traits

AggregateAuthenticator

Trait impl’d by aggregated signatures in asymmetric cryptography.

AuthenticatedCipher

Trait impl’d by symmetric ciphers for authenticated encryption.

Authenticator

Trait impl’d by signatures in asymmetric cryptography.

Cipher

Trait impl’d by symmetric ciphers.

EncodeDecodeBase64

Cryptographic material with an immediate conversion to/from Base64 strings.

EncryptionKey

Trait impl’d by encryption keys in symmetric cryptography

FromUniformBytes

Trait impl’d by a keys/secret seeds for generating a secure instance.

Generate

Trait impl’d by cryptographic material that can be generated randomly such as keys and nonces.

KeyPair

Trait impl’d by a public / private key pair in asymmetric cryptography.

Nonce

Trait impl’d by nonces and IV’s used in symmetric cryptography

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

SigningKey

Trait impl’d by private (secret) keys in asymmetric cryptography.

ToFromBytes

Trait impl’d by concrete types that represent digital cryptographic material (keys). For signatures, we rely on signature::Signature, which may be more widely implemented.

VerifyingKey

Trait impl’d by public keys in asymmetric cryptography.