Trait fastcrypto::traits::AggregateAuthenticator

pub trait AggregateAuthenticator: Display + Default + Serialize + DeserializeOwned + Send + Sync + 'static + Clone {
    type Sig: Authenticator<PubKey = Self::PubKey>;
    type PubKey: VerifyingKey<Sig = Self::Sig>;
    type PrivKey: SigningKey<Sig = Self::Sig>;

    fn aggregate<'a, K: Borrow<Self::Sig> + 'a, I: IntoIterator<Item = &'a K>>(
        signatures: I
    ) -> Result<Self, FastCryptoError>;
    fn add_signature(
        &mut self,
        signature: Self::Sig
    ) -> Result<(), FastCryptoError>;
    fn add_aggregate(&mut self, signature: Self) -> Result<(), FastCryptoError>;
    fn verify(
        &self,
        pks: &[<Self::Sig as Authenticator>::PubKey],
        message: &[u8]
    ) -> Result<(), FastCryptoError>;
    fn verify_different_msg(
        &self,
        pks: &[<Self::Sig as Authenticator>::PubKey],
        messages: &[&[u8]]
    ) -> Result<(), FastCryptoError>;
    fn batch_verify<'a>(
        sigs: &[&Self],
        pks: Vec<impl ExactSizeIterator<Item = &'a Self::PubKey>>,
        messages: &[&[u8]]
    ) -> Result<(), FastCryptoError>;
}

Trait impl’d by aggregated signatures in asymmetric cryptography.

The trait bounds are implemented to allow the aggregation of multiple signatures, and to verify it against multiple, unaggregated public keys. For signature schemes where aggregation is not possible, a trivial implementation is provided.

Required Associated Types

type Sig: Authenticator<PubKey = Self::PubKey>

type PubKey: VerifyingKey<Sig = Self::Sig>

type PrivKey: SigningKey<Sig = Self::Sig>

Required Methods

fn aggregate<'a, K: Borrow<Self::Sig> + 'a, I: IntoIterator<Item = &'a K>>(
signatures: I
) -> Result<Self, FastCryptoError>

Combine signatures into a single aggregated signature.

fn add_signature(&mut self, signature: Self::Sig) -> Result<(), FastCryptoError>

fn add_aggregate(&mut self, signature: Self) -> Result<(), FastCryptoError>

fn verify(
&self,
pks: &[<Self::Sig as Authenticator>::PubKey],
message: &[u8]
) -> Result<(), FastCryptoError>

fn verify_different_msg(
&self,
pks: &[<Self::Sig as Authenticator>::PubKey],
messages: &[&[u8]]
) -> Result<(), FastCryptoError>

fn batch_verify<'a>(
sigs: &[&Self],
pks: Vec<impl ExactSizeIterator<Item = &'a Self::PubKey>>,
messages: &[&[u8]]
) -> Result<(), FastCryptoError>