Trait fastcrypto::traits::AuthenticatedCipher

pub trait AuthenticatedCipher {
    type IVType: Nonce;

    fn encrypt_authenticated(
        &self,
        iv: &Self::IVType,
        aad: &[u8],
        plaintext: &[u8]
    ) -> Result<Vec<u8>, FastCryptoError>;
    fn decrypt_authenticated(
        &self,
        iv: &Self::IVType,
        aad: &[u8],
        ciphertext: &[u8]
    ) -> Result<Vec<u8>, FastCryptoError>;
}

Trait impl’d by symmetric ciphers for authenticated encryption.

Required Associated Types

type IVType: Nonce

Required Methods

fn encrypt_authenticated(
&self,
iv: &Self::IVType,
aad: &[u8],
plaintext: &[u8]
) -> Result<Vec<u8>, FastCryptoError>

Encrypt plaintext and write result to buffer using the given IV and authentication data

fn decrypt_authenticated(
&self,
iv: &Self::IVType,
aad: &[u8],
ciphertext: &[u8]
) -> Result<Vec<u8>, FastCryptoError>

Decrypt ciphertext and write result to buffer using the given IV and authentication data