Trait fastcrypto::traits::Cipher

pub trait Cipher {
    type IVType: Nonce;

    fn encrypt(
        &self,
        iv: &Self::IVType,
        plaintext: &[u8]
    ) -> Result<Vec<u8>, FastCryptoError>;
    fn decrypt(
        &self,
        iv: &Self::IVType,
        ciphertext: &[u8]
    ) -> Result<Vec<u8>, FastCryptoError>;
}

Trait impl’d by symmetric ciphers.

Required Associated Types

type IVType: Nonce

Required Methods

fn encrypt(
&self,
iv: &Self::IVType,
plaintext: &[u8]
) -> Result<Vec<u8>, FastCryptoError>

Encrypt plaintext and write result to buffer using the given IV

fn decrypt(
&self,
iv: &Self::IVType,
ciphertext: &[u8]
) -> Result<Vec<u8>, FastCryptoError>

Decrypt ciphertext and write result to buffer using the given IV