fastcrypto::traits

Trait InsecureDefault

pub trait InsecureDefault {
    // Required method
    fn insecure_default() -> Self;
}

Trait for objects that support an insecure default value that should only be used as a placeholder.

Required Methods

fn insecure_default() -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.