Trait fastcrypto::traits::VerifyingKey

pub trait VerifyingKey: Serialize + DeserializeOwned + Hash + Display + Eq + Ord + Default + ToFromBytes + Verifier<Self::Sig> + for<'a> From<&'a Self::PrivKey> + Send + Sync + 'static + Clone {
    type PrivKey: SigningKey<PubKey = Self>;
    type Sig: Authenticator<PubKey = Self>;

    const LENGTH: usize;

    fn verify_batch_empty_fail(
        msg: &[u8],
        pks: &[Self],
        sigs: &[Self::Sig]
    ) -> Result<(), Report> { ... }
    fn verify_batch_empty_fail_different_msg<'a, M>(
        msgs: &[M],
        pks: &[Self],
        sigs: &[Self::Sig]
    ) -> Result<(), Report>
    where
        M: Borrow<[u8]> + 'a,
    { ... }
}

Trait impl’d by public keys in asymmetric cryptography.

The trait bounds are implemented so as to be symmetric and equivalent to the ones on its associated types for private and signature material.

Required Associated Types

type PrivKey: SigningKey<PubKey = Self>

type Sig: Authenticator<PubKey = Self>

Required Associated Constants

const LENGTH: usize

Provided Methods

fn verify_batch_empty_fail(
msg: &[u8],
pks: &[Self],
sigs: &[Self::Sig]
) -> Result<(), Report>

Batch verification over the same message. Implementations of this method can be fast, assuming rogue key checks have already been performed. TODO: take as input a flag to denote if rogue key protection already took place.

Example

use fastcrypto::ed25519::*;
use rand::thread_rng;
let message: &[u8] = b"Hello, world!";
let kp1 = Ed25519KeyPair::generate(&mut thread_rng());
let signature1 = kp1.sign(message);
let kp2 = Ed25519KeyPair::generate(&mut thread_rng());
let signature2 = kp2.sign(message);
let public_keys = [kp1.public().clone(), kp2.public().clone()];
let signatures = [signature1.clone(), signature2.clone()];
assert!(Ed25519PublicKey::verify_batch_empty_fail(message, &public_keys, &signatures).is_ok());

fn verify_batch_empty_fail_different_msg<'a, M>(
msgs: &[M],
pks: &[Self],
sigs: &[Self::Sig]
) -> Result<(), Report>where
M: Borrow<[u8]> + 'a,

Batch verification over different messages. Implementations of this method can be fast, assuming rogue key checks have already been performed. TODO: take as input a flag to denote if rogue key protection already took place.

Example

use fastcrypto::ed25519::*;
use rand::thread_rng;
let message1: &[u8] = b"Hello, world!";
let kp1 = Ed25519KeyPair::generate(&mut thread_rng());
let signature1 = kp1.sign(message1);
let message2: &[u8] = b"Hello, world!!!";
let kp2 = Ed25519KeyPair::generate(&mut thread_rng());
let signature2 = kp2.sign(message2);
let messages = [message1, message2];
let public_keys = [kp1.public().clone(), kp2.public().clone()];
let signatures = [signature1.clone(), signature2.clone()];
assert!(Ed25519PublicKey::verify_batch_empty_fail_different_msg(&messages, &public_keys, &signatures).is_ok());