fastcrypto::vrf

Trait VRFPublicKey

pub trait VRFPublicKey {
    type PrivateKey: VRFPrivateKey<PublicKey = Self>;
}

Represents a public key of which is use to verify outputs for a verifiable random function (VRF).

Required Associated Types

type PrivateKey: VRFPrivateKey<PublicKey = Self>