Crate fastcrypto

Re-exports

pub use signature::Signature as _;

Modules

aes

This module contains implementations of various AES modes.

bls12377

bls12381

bulletproofs

ed25519

This module contains an implementation of the Ed25519 signature scheme.

encoding

error

hash

This module contains a selection of cryptographic hash functions implementing a common HashFunction trait.

hmac

private_seed

pubkey_bytes

secp256k1

serde_helpers

traits

Structs

SignatureService

This service holds the node’s private key. It takes digests as input and returns a signature over the digest (through a one-shot channel).

Traits

Verifier

Verify the provided message bytestring using Self (e.g. a public key)

Functions

generate_keypair

generate_production_keypair