Crate fastcrypto

Re-exports

pub use signature::Signature as _;

Modules

aes

This module contains implementations of various AES modes.

bls12377

This module contains an implementation of the BLS signature scheme over the BLS 12-377 curve.

bls12381

This module contains an implementation of the BLS signature scheme over the BLS 12-381 curve.

bulletproofs

Implementation of Pedersen Commitments and Bulletproofs, which are short non-interactive zero-knowledge proofs that require no trusted setup.

ed25519

This module contains an implementation of the Ed25519 signature scheme.

encoding

Encodings of binary data such as Base64 and Hex.

error

Collection of errors to be used in fastcrypto.

hash

This module contains a selection of cryptographic hash functions implementing a common HashFunction trait.

hmac

Implementations of HMAC and HKDF.

private_seed

pubkey_bytes

secp256k1

This module contains an implementation of the ECDSA signature scheme over the secp256k1 curve.

serde_helpers

traits

Structs

SignatureService

This service holds the node’s private key. It takes digests as input and returns a signature over the digest (through a one-shot channel).

Traits

Verifier

Verify the provided message bytestring using Self (e.g. a public key)

Functions

generate_keypair

generate_production_keypair