Crate fastcrypto

Modules

aes

This module contains implementations of various AES modes.

bls12381

This module contains an implementation of the BLS signature scheme over the BLS 12-381 curve.

bulletproofs

Implementation of Pedersen Commitments and Bulletproofs, which are short non-interactive zero-knowledge proofs that require no trusted setup.

ed25519

This module contains an implementation of the Ed25519 signature scheme.

encoding

Encodings of binary data such as Base64 and Hex.

error

Collection of errors to be used in fastcrypto.

groups

hash

This module contains a selection of cryptographic hash functions implementing a common HashFunction trait.

hmac

Implementations of HMAC and HKDF.

jwt_utils

merkle

A simple Merkle tree implementation.

nizk

pedersen

private_seed

random_oracle

rsa

Implementation of a verifier following RSASSA-PKCS1-v1_5 using SHA-256 (see https://datatracker.ietf.org/doc/rfc3447/).

secp256k1

This module contains an implementation of the ECDSA signature scheme over the secp256k1 curve.

secp256r1

This module contains an implementation of the ECDSA signature scheme over the secp256r1 NIST-P1 curve. The nonce is generated deterministically according to RFC6979.

serde_helpers

signature_service

traits

twisted_elgamal

utils

vrf

Macros

generate_bytes_representation

Macro for generating a new alias for BytesRepresentation with the given $length, and From functions for both directions.

impl_base64_display_fmt

serialize_deserialize_with_to_from_byte_array

Macro for generating Serialize/Deserialize for a type that implements ToFromByteArray.

serialize_deserialize_with_to_from_bytes