fastcrypto

Module bls12381

This module contains an implementation of the BLS signature scheme over the BLS 12-381 curve.

use rand::thread_rng;
let kp = BLS12381KeyPair::generate(&mut thread_rng());
let message: &[u8] = b"Hello, world!";
let signature = kp.sign(message);
assert!(kp.public().verify(message, &signature).is_ok());

Modules

min_pk

Module minimizing the size of public keys. Module minimizing the size of public keys. See also min_sig.

min_sig

Module minimizing the size of signatures. Module minimizing the size of signatures. See also min_pk.

Constants

BLS_G1_LENGTH

The length of public keys when using the min_pk module and the length of signatures when using the min_sig module.

BLS_G2_LENGTH

The length of public keys when using the min_sig module and the length of signatures when using the min_pk module.

BLS_KEYPAIR_LENGTH

The key pair bytes length used by helper is the same as the private key length. This is because only private key is serialized.

BLS_PRIVATE_KEY_LENGTH

The length of a private key in bytes.