fastcrypto::groups

Module bls12381

Structs

G1Element

Elements of the group G_1 in BLS 12-381.

G1ElementUncompressed

An uncompressed serialization of a G1 element. This format is two times longer than the compressed format used by G1Element::serialize, but is much faster to deserialize.

G2Element

Elements of the group G_2 in BLS 12-381.

GTElement

Elements of the subgroup G_T of F_q^{12} in BLS 12-381. Note that it is written in additive notation here.

Scalar

This represents a scalar modulo r = 0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001 which is the order of the groups G1, G2 and GT. Note that r is a 255 bit prime.

Constants

FP_BYTE_LENGTH

G1_ELEMENT_BYTE_LENGTH

G2_ELEMENT_BYTE_LENGTH

GT_ELEMENT_BYTE_LENGTH

SCALAR_LENGTH

Functions

buffer_to_scalar_mod_r

Similar to reduce_mod_uniform_buffer, returns a result of scalar, and does not panic on invalid length.

Type Aliases

G1ElementAsBytes

G2ElementAsBytes