Trait fastcrypto::groups::multiplier::ScalarMultiplier

pub trait ScalarMultiplier<G, S> {
    // Required methods
    fn new(base_element: G, zero: G) -> Self;
    fn mul(&self, scalar: &S) -> G;
    fn two_scalar_mul(
        &self,
        base_scalar: &S,
        other_element: &G,
        other_scalar: &S
    ) -> G;
}

Trait for scalar multiplication for a fixed group element, e.g. by using precomputed values.

Required Methods

fn new(base_element: G, zero: G) -> Self

Create a new scalar multiplier with the given base element.

fn mul(&self, scalar: &S) -> G

Compute self.base_element * scalar.

fn two_scalar_mul( &self, base_scalar: &S, other_element: &G, other_scalar: &S ) -> G

Compute self.base_element * base_scalar + other_element * other_scalar.

Object Safety

This trait is not object safe.