fastcrypto::groups

Module ristretto255

Implementations of the ristretto255 group which is a group of prime order 2^{252} + 27742317777372353535851937790883648493 built over Curve25519.

Structs

RistrettoPoint

Represents a point in the Ristretto group for Curve25519.

RistrettoScalar

Represents a scalar.