Trait fastcrypto::groups::GroupElement

pub trait GroupElement: Copy + Clone + Eq + Add<Output = Self> + AddAssign + for<'a> Add<&'a Self, Output = Self> + Sub<Output = Self> + SubAssign + for<'a> Sub<&'a Self, Output = Self> + Neg<Output = Self> + Mul<Self::ScalarType, Output = Self> + for<'a> Mul<&'a Self::ScalarType, Output = Self> + Sized {
    type ScalarType: Scalar;

    fn zero() -> Self;
    fn generator() -> Self;
}

Trait impl’d by elements of an additive cyclic group.

Required Associated Types

type ScalarType: Scalar

Type of scalars used in the [Self::mul] multiplication method.

Required Methods

fn zero() -> Self

Return an instance of the identity element in this group.

fn generator() -> Self

Return an instance of the generator for this group.