Module fastcrypto::hash

This module contains a selection of cryptographic hash functions implementing a common HashFunction trait.

Example

let digest1 = Sha256::digest(b"Hello, world!");

let mut hash_function = Sha256::default();
hash_function.update(b"Hello, ");
hash_function.update(b"world!");
let digest2 = hash_function.finalize();

assert_eq!(digest1, digest2);

Structs

Blake3

The BLAKE3 hash function with 256 bit digests.

Digest

Represents a digest of DIGEST_LEN bytes.

HashFunctionWrapper

This wraps a [digest::Digest] as a HashFunction.

Traits

Hash

This trait is implemented by all messages that can be hashed.

HashFunction

Trait implemented by hash functions providing a output of fixed length

ReverseWrapper

This trait allows using a HashFunctionWrapper where a [digest::Digest] was expected.

Type Definitions

Blake2b256

The BLAKE2-256 hash function with 256 bit digests.

Keccak256

The KECCAK hash function with 256 bit digests.

Sha3_256

The SHA-3 hash function with 256 bit digests.

Sha256

The SHA-2 hash function with 256 bit digests.

Sha512

The SHA-512 hash function with 512 bit digests.