Module fastcrypto::hash

Structs

Blake3

BLAKE3

Digest

Represents a digest of DIGEST_LENbytes.

HashFunctionWrapper

This wraps a digest::Digest as a fastcrypto::hash::HashFunction.

Traits

Hash

This trait is implemented by all messages that can be hashed.

HashFunction

Trait implemented by hash functions providing a output of fixed length

ReverseWrapper

This allows us to use the hash functions defined here and use them in place of their internal functions from the digest crate.

Type Definitions

Blake2b256

BLAKE2-256

Keccak256

KECCAK

Sha3_256

SHA-3

Sha256

SHA-2

Sha512

SHA-512