Trait fastcrypto::hash::HashFunction

pub trait HashFunction<DigestLength: ArrayLength<u8>>: OutputSizeUser + Sized + Default {
    fn update(&mut self, data: &[u8]);
    fn finalize(self) -> Digest<DigestLength>;

    fn digest(data: &[u8]) -> Digest<DigestLength> { ... }
}

Trait implemented by hash functions providing a output of fixed length

Required Methods

fn update(&mut self, data: &[u8])

Process the given data, and update the internal of the hash function.

fn finalize(self) -> Digest<DigestLength>

Retrieve result and consume hash function.

Provided Methods

fn digest(data: &[u8]) -> Digest<DigestLength>