Trait fastcrypto::hash::MultisetHash

pub trait MultisetHash<I>: Eq {
    fn insert(&mut self, item: &I);
    fn insert_all<'a, It>(&'a mut self, items: It)
    where
        It: 'a + IntoIterator<Item = &'a I>,
        I: 'a;
    fn union(&mut self, other: &Self);
}

A Multiset Hash is a homomorphic hash function, which hashes arbitrary multisets of objects such that the hash of the union of two multisets is easy to compute from the hashes of the two multisets.

The hash may be computed incrementally, adding items one at a time, and the order does not affect the result. The hash of two multisets can be compared by using the Eq trait impl’d for the given hash function, and the hash function should be collision resistant.

See “Incremental Multiset Hash Functions and Their Application to Memory Integrity Checking” by D. Clarke et al. for a discussion of this type of hash functions.

Example

use fastcrypto::hash::{EllipticCurveMultisetHash, MultisetHash};

let mut hash1 = EllipticCurveMultisetHash::default();
hash1.insert(b"Hello");
hash1.insert(b"World");

let mut hash2 = EllipticCurveMultisetHash::default();
hash2.insert(b"World");
hash2.insert(b"Hello");

assert_eq!(hash1, hash2);

Required Methods

fn insert(&mut self, item: &I)

Insert an item into this hash function.

fn insert_all<'a, It>(&'a mut self, items: It)where
It: 'a + IntoIterator<Item = &'a I>,
I: 'a,

Insert multiple items into this hash function.

fn union(&mut self, other: &Self)

Add all the elements of another hash function into this hash function.