Function fastcrypto::hmac::hkdf_generate_from_ikm

pub fn hkdf_generate_from_ikm<H, K>(
    ikm: &[u8],
    salt: &[u8],
    info: &[u8]
) -> Result<K, FastCryptoError>where
    H: ReverseWrapper,
    <<H as ReverseWrapper>::Variant as CoreProxy>::Core: HashMarker + UpdateCore + FixedOutputCore + BufferKindUser<BufferKind = Eager> + Default + Clone,
    <<<H as ReverseWrapper>::Variant as CoreProxy>::Core as BlockSizeUser>::BlockSize: IsLess<U256>,
    Le<<<<H as ReverseWrapper>::Variant as CoreProxy>::Core as BlockSizeUser>::BlockSize, U256>: NonZero,
    K: KeyPair,

Creation of a keypair using the RFC 5869 HKDF specification. This requires choosing an HMAC function of the correct length (conservatively, the size of a private key for this curve). Despite the unsightly generics (which aim to ensure this works for a wide range of hash functions), this is straightforward to use.

Example:

use fastcrypto::hash::Sha3_256;
use fastcrypto::ed25519::Ed25519KeyPair;
use fastcrypto::hmac::hkdf_generate_from_ikm;
    let ikm = b"some_ikm";
    let info = b"my_app";
    let salt = b"some_salt";
    let my_keypair = hkdf_generate_from_ikm::<Sha3_256, Ed25519KeyPair>(ikm, salt, info);

    let my_keypair_default_info = hkdf_generate_from_ikm::<Sha3_256, Ed25519KeyPair>(ikm, salt, &[]);

Note: This HKDF function may not match the native library’s deterministic key generation functions. For example, observe that in the blst library:

use fastcrypto::bls12381::min_sig::BLS12381KeyPair;
use fastcrypto::hash::Sha3_256;
use fastcrypto::traits::{KeyPair, SigningKey, ToFromBytes};
use fastcrypto::hmac::hkdf_generate_from_ikm;

    let ikm = b"02345678001234567890123456789012";
    let info = b"my_app";
    let salt = b"some_salt";
    let my_keypair = hkdf_generate_from_ikm::<Sha3_256, BLS12381KeyPair>(ikm, salt, info).unwrap();
    let native_sk = blst::min_sig::SecretKey::key_gen_v4_5(ikm, salt, info).unwrap();
    assert_ne!(native_sk.to_bytes(), my_keypair.private().as_bytes());