Module fastcrypto::secp256k1

This module contains an implementation of the ECDSA signature scheme over the secp256k1 curve.

Messages can be signed and the signature can be verified again:

Example

use rand::thread_rng;
let kp = Secp256k1KeyPair::generate(&mut thread_rng());
let message: &[u8] = b"Hello, world!";
let signature = kp.sign(message);
assert!(kp.public().verify(message, &signature).is_ok());

Structs

Secp256k1KeyPair

Secp256k1 public/private key pair.

Secp256k1PrivateKey

Secp256k1 private key.

Secp256k1PublicKey

Secp256k1 public key.

Secp256k1Signature

Secp256k1 signature.

Constants

RECOVERABLE_SIGNATURE_SIZE

Length of a compact signature followed by one extra byte for recovery id, used to recover the public key from a signature.

Statics

SECP256K1

Type Definitions

Secp256k1PublicKeyBytes

Binary representation of an instance of Secp256k1PublicKey.