Module fastcrypto::secp256r1

This module contains an implementation of the ECDSA signature scheme over the secp256r1 NIST-P1 curve. The nonce is generated deterministically according to RFC6979.

Messages can be signed and the signature can be verified again:

Example

use rand::thread_rng;
let kp = Secp256r1KeyPair::generate(&mut thread_rng());
let message: &[u8] = b"Hello, world!";
let signature = kp.sign(message);
assert!(kp.public().verify(message, &signature).is_ok());

Modules

recoverable

This module contains an implementation of the ECDSA signature scheme over the secp256r1 NIST-P1 curve where the public key can be recovered from a signature. The nonce is generated deterministically according to RFC6979.

Structs

Secp256r1KeyPair

Secp256r1 public/private key pair.

Secp256r1PrivateKey

Secp256r1 private key.

Secp256r1PublicKey

Secp256r1 public key.

Secp256r1Signature

Secp256r1 signature.

Constants

SECP256R1_KEYPAIR_LENGTH

The key pair bytes length is the same as the private key length. This enforces deserialization to always derive the public key from the private key.

SECP256R1_PRIVATE_KEY_LENGTH

The length of a private key in bytes.

SECP256R1_PUBLIC_KEY_LENGTH

The length of a public key in bytes.

SECP256R1_SIGNATURE_LENTH

The length of a signature in bytes.