fastcrypto

Module serde_helpers

Structs

BytesRepresentation

External types.

SerializationHelper

Serialization of internal types.

Traits

ToFromByteArray

Functions

deserialize_vector

Given a byte array of length N * SIZE_IN_BYTES, deserialize it into a vector of N elements of type T with the given deserialization function.

keypair_decode_base64

serialize_vector

Serialize a vector of elements of type T into a byte array by simply concatenating their binary representations with the given deserialization function.