fastcrypto
Module traits

Traits

AggregateAuthenticator

Trait impl’d by aggregated signatures in asymmetric cryptography.

AllowedRng

Trait impl’d by RNG’s accepted by fastcrypto.

Authenticator

Trait impl’d by signatures in asymmetric cryptography.

EncodeDecodeBase64

Cryptographic material with an immediate conversion to/from Base64 strings.

FromUniformBytes

Trait impl’d by a keys/secret seeds for generating a secure instance.

Generate

Trait impl’d by cryptographic material that can be generated randomly such as keys and nonces.

InsecureDefault

Trait for objects that support an insecure default value that should only be used as a placeholder.

KeyPair

Trait impl’d by a public / private key pair in asymmetric cryptography.

RecoverableSignature

Trait impl’d by recoverable signatures

RecoverableSigner

Trait impl’d by public / private keypairs that can generate recoverable signatures

Signer

Trait impl’d by a key/keypair that can create signatures.

SigningKey

Trait impl’d by private (secret) keys in asymmetric cryptography.

ToFromBytes

Trait impl’d by concrete types that represent digital cryptographic material (keys).

VerifyRecoverable

VerifyingKey

Trait impl’d by public keys in asymmetric cryptography.