fastcrypto::traits

Trait Generate

pub trait Generate {
    // Required method
    fn generate<R: AllowedRng>(rng: &mut R) -> Self;
}

Trait impl’d by cryptographic material that can be generated randomly such as keys and nonces.

Required Methods

fn generate<R: AllowedRng>(rng: &mut R) -> Self

Generate a new random instance using the given RNG.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.