Trait fastcrypto::traits::InsecureDefault

pub trait InsecureDefault {
    // Required method
    fn insecure_default() -> Self;
}

Trait for objects that support an insecure default value that should only be used as a placeholder.

Required Methods

fn insecure_default() -> Self