fastcrypto::traits

Trait SigningKey

pub trait SigningKey:
    ToFromBytes
    + Serialize
    + DeserializeOwned
    + Send
    + Sync
    + 'static {
    type PubKey: VerifyingKey<PrivKey = Self>;
    type Sig: Authenticator<PrivKey = Self>;

    const LENGTH: usize;
}

Trait impl’d by private (secret) keys in asymmetric cryptography.

The trait bounds are implemented so as to be symmetric and equivalent to the ones on its associated types for public key and signature material.

Required Associated Constants

const LENGTH: usize

Required Associated Types

type PubKey: VerifyingKey<PrivKey = Self>

type Sig: Authenticator<PrivKey = Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.