Module fastcrypto::vrf

Modules

ecvrf

An implementation of an Elliptic Curve VRF (ECVRF) using the Ristretto255 group. The implementation follows the specifications in draft-irtf-cfrg-vrf-15 (https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/).

Traits

VRFKeyPair

A keypair for a verifiable random function (VRF).

VRFPrivateKey

Represents a private key used to compute outputs for a verifiable random function (VRF).

VRFProof

A proof that the output of a VRF was computed correctly.

VRFPublicKey

Represents a public key of which is use to verify outputs for a verifiable random function (VRF).